Remove grpc server from public api
