+++ projects/services/public-api-service-typescript/src/index.ts
@@ -4,6 +4,7 @@ import { ErrorHandler, grpc, Service, ServiceRunner, topology, topologyPeers, lo
 
 import PublicApiService from "./service";
 import PublicApiHTTPService from "./http-service";
+import httpServer from "./server";
 
 const { NODE_NAME } = process.env;
 
@@ -11,22 +12,5 @@ ErrorHandler.setup();
 
 Service.initLogger(path.join(__dirname, "../../../../logs/public-api.log"));
 
-if (!NODE_NAME) {
-  throw new Error("NODE_NAME can't be empty!");
-}
-
 const nodeTopology = topology();
-const peers = topologyPeers(nodeTopology.peers);
-const nodeConfig = {
-  nodeName: NODE_NAME
-};
-
-ServiceRunner.run(grpc.publicApiServer, new PublicApiService(peers.virtualMachine, peers.transactionPool, nodeConfig), nodeTopology.endpoint);
-
-const httpNodeConfig = {
-  nodeName: NODE_NAME,
-  httpPort: 80
-};
-
-const httpService = new PublicApiHTTPService(peers.virtualMachine, peers.transactionPool, httpNodeConfig);
-httpService.start();
+httpServer(nodeTopology, process.env).start();
